ars-agents/app/requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain-core
langchain-openai
langchain-community
pydantic
openai
requests
beautifulsoup4